Improve security
- Refresh Tokens should be stored securely, The
token is essentially the user's password. It should be stored in the
safest place possible in the browser which is a Secure/httpOnly cookie
where it can't be read/manipulated from JavaScript

- Ensure that the user presents an expired token before granting them a
new one. If a refreshtoken is leaked, the attacker must present an
expired JWT alongside the refresh token before they receive a new token.
This adds an extra layer of protection.